aws-bedrock-a2a-proxy/pyproject.toml
[project]
name = "aws-bedrock-a2a-proxy"
version = "0.1.0"
description = "A2A server for AWS Bedrock AgentCore agents"
authors = [
    {name = "Dave Kerr", email = "dave@example.com"}
]
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.34.2",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "a2a-sdk[sqlite]>=0.2.12",
    "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.2",
    "mypy>=1.11.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
aws-bedrock-a2a-proxy = "aws_bedrock_a2a_proxy.__main__:main"

[tool.ruff]
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["aws_bedrock_a2a_proxy"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
warn_return_any = true
warn_unused_configs = true
no_implicit_optional = true